fix: remove unnecessary View, move constants and styles to new files, add prop types to DayPicker
