+++ source/components/molecules/DayPicker/localeConfig.js
@@ -1,0 +1,34 @@
+/* passed to LocaleConfig['se'] in `react-native-calendars` in order to localize calendar text */
+export default {
+  monthNames: [
+    'januari',
+    'februari',
+    'mars',
+    'april',
+    'maj',
+    'juni',
+    'juli',
+    'augusti',
+    'september',
+    'october',
+    'november',
+    'december',
+  ],
+  monthNamesShort: [
+    'jan',
+    'feb',
+    'mar',
+    'apr',
+    'maj',
+    'jun',
+    'jul',
+    'aug',
+    'sep',
+    'oct',
+    'nov',
+    'dec',
+  ],
+  dayNames: ['söndag', 'måndag', 'tisdag', 'onsdag', 'torsdag', 'fredag', 'lördag'],
+  dayNamesShort: ['sön', 'mån', 'tis', 'ons', 'tor', 'fre', 'mån'],
+  today: 'idag',
+};
